mkfs.cmfs: change VERSION to string format.
In mkfs.cmfs/Makefile.am, VERSION was defined as -DVERSION=$(VERSION).
When VERSION is defined as major.minor, e.g. 0.1, macro VERSION will
be assigned as a double float number.

This patch changes the definition to -DVERSION=\"$(VERSION)\", so
macro VERSION will be assigned as a string format. Gcc will not
complain when we call printf with %s format string to display VERSION
string.